// src/integrations/supabase/types.d.ts
export type Database = {
  public: {
    Tables: {
// ...
      contact_submissions: {
        Row: {
          checkbox: boolean | null
          contacted: boolean | null
          created_at: string
          email: string
          id: string
          message: string
          name: string
          phone: string
        }
        Insert: {
          checkbox?: boolean | null
          contacted?: boolean | null
          created_at?: string
          email: string
          id?: string
          message: string
          name: string
          phone: string
        }
        Update: {
          checkbox?: boolean | null
          contacted?: boolean | null
          created_at?: string
          email?: string
          id?: string
          message?: string
          name?: string
          phone?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}